+++ scompose/utils/__init__.py
@@ -1,21 +1,12 @@
-'''
+"""
 
-Copyright (C) 2019 Vanessa Sochat.
+Copyright (C) 2019-2020 Vanessa Sochat.
 
-This program is free software: you can redistribute it and/or modify it
-under the terms of the GNU Affero General Public License as published by
-the Free Software Foundation, either version 3 of the License, or (at your
-option) any later version.
+This Source Code Form is subject to the terms of the
+Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
+with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-This program is distributed in the hope that it will be useful, but WITHOUT
-ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public
-License for more details.
-
-You should have received a copy of the GNU Affero General Public License
-along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-'''
+"""
 
 import errno
 import json
@@ -25,26 +16,23 @@ import re
 import sys
 import yaml
 
-from subprocess import (
-    Popen,
-    PIPE,
-    STDOUT
-)
+from subprocess import Popen, PIPE, STDOUT
 
 
 def get_installdir():
-    '''get_installdir returns the installation directory of the application
-    '''
+    """get_installdir returns the installation directory of the application
+    """
     return os.path.abspath(os.path.dirname(os.path.dirname(__file__)))
 
 
 def get_userhome():
-    '''get the user home based on the effective uid
-    '''
+    """get the user home based on the effective uid
+    """
     return pwd.getpwuid(os.getuid())[5]
 
+
 def run_command(cmd, sudo=False):
-    '''run_command uses subprocess to send a command to the terminal.
+    """run_command uses subprocess to send a command to the terminal.
 
         Parameters
         ==========
@@ -52,9 +40,9 @@ def run_command(cmd, sudo=False):
         error_message: the error message to give to user if fails,
         if none specified, will alert that command failed.
 
-    '''
+    """
     if sudo is True:
-        cmd = ['sudo'] + cmd
+        cmd = ["sudo"] + cmd
 
     try:
         output = Popen(cmd, stderr=STDOUT, stdout=PIPE)
@@ -64,11 +52,10 @@ def run_command(cmd, sudo=False):
         output = Popen(cmd, stderr=STDOUT, stdout=PIPE)
 
     t = output.communicate()[0], output.returncode
-    output = {'message': t[0],
-              'return_code': t[1]}
+    output = {"message": t[0], "return_code": t[1]}
 
-    if isinstance(output['message'], bytes):
-        output['message'] = output['message'].decode('utf-8')
+    if isinstance(output["message"], bytes):
+        output["message"] = output["message"].decode("utf-8")
 
     return output
 
@@ -79,9 +66,9 @@ def run_command(cmd, sudo=False):
 
 
 def mkdir_p(path):
-    '''mkdir_p attempts to get the same functionality as mkdir -p
+    """mkdir_p attempts to get the same functionality as mkdir -p
     :param path: the path to create.
-    '''
+    """
     try:
         os.makedirs(path)
     except OSError as e:
@@ -98,18 +85,18 @@ def mkdir_p(path):
 
 
 def write_file(filename, content, mode="w"):
-    '''write_file will open a file, "filename" and write content, "content"
+    """write_file will open a file, "filename" and write content, "content"
        and properly close the file
-    '''
+    """
     with open(filename, mode) as filey:
         filey.writelines(content)
     return filename
 
 
 def read_file(filename, mode="r", readlines=True):
-    '''write_file will open a file, "filename" and write content, "content"
+    """write_file will open a file, "filename" and write content, "content"
        and properly close the file
-    '''
+    """
     with open(filename, mode) as filey:
         if readlines is True:
             content = filey.readlines()
@@ -117,37 +104,39 @@ def read_file(filename, mode="r", readlines=True):
             content = filey.read()
     return content
 
+
 # Yaml
 
-def read_yaml(filename, mode='r', quiet=False):
-    '''read a yaml file, only including sections between dashes
-    '''
+
+def read_yaml(filename, mode="r", quiet=False):
+    """read a yaml file, only including sections between dashes
+    """
     stream = read_file(filename, mode, readlines=False)
     return _read_yaml(stream, quiet=quiet)
 
 
 def write_yaml(yaml_dict, filename, mode="w"):
-    '''write a dictionary to yaml file
+    """write a dictionary to yaml file
  
        Parameters
        ==========
        yaml_dict: the dict to print to yaml
        filename: the output file to write to
        pretty_print: if True, will use nicer formatting
-    '''
+    """
     with open(filename, mode) as filey:
         filey.writelines(yaml.dump(yaml_dict))
     return filename
 
-   
+
 def _read_yaml(section, quiet=False):
-    '''read yaml from a string, either read from file (read_frontmatter) or 
+    """read yaml from a string, either read from file (read_frontmatter) or 
        from yml file proper (read_yaml)
 
        Parameters
        ==========
        section: a string of unparsed yaml content.
-    '''
+    """
     metadata = {}
 
     # PyYaml vs pyaml have subtle differences
@@ -160,22 +149,23 @@ def _read_yaml(section, quiet=False):
         if isinstance(doc, dict):
             for k, v in doc.items():
                 if not quiet:
-                    print('%s: %s' %(k, v))
+                    print("%s: %s" % (k, v))
                 metadata[k] = v
     return metadata
 
 
 # Json
 
+
 def write_json(json_obj, filename, mode="w", print_pretty=True):
-    '''write_json will (optionally,pretty print) a json object to file
+    """write_json will (optionally,pretty print) a json object to file
 
        Parameters
        ==========
        json_obj: the dict to print to json
        filename: the output file to write to
        pretty_print: if True, will use nicer formatting
-    '''
+    """
     with open(filename, mode) as filey:
         if print_pretty:
             filey.writelines(print_json(json_obj))
@@ -185,19 +175,15 @@ def write_json(json_obj, filename, mode="w", print_pretty=True):
 
 
 def print_json(json_obj):
-    ''' just dump the json in a "pretty print" format
-    '''
-    return json.dumps(
-                    json_obj,
-                    indent=4,
-                    separators=(
-                        ',',
-                        ': '))
+    """ just dump the json in a "pretty print" format
+    """
+    return json.dumps(json_obj, indent=4, separators=(",", ": "))
 
-def read_json(filename, mode='r'):
-    '''read_json reads in a json file and returns
+
+def read_json(filename, mode="r"):
+    """read_json reads in a json file and returns
        the data structure as dict.
-    '''
+    """
     with open(filename, mode) as filey:
         data = json.load(filey)
     return data